reorder pickle-able helpers
